+++ gradle-plugins/src/main/kotlin/tz/co/asoft/PrepareWebpackConfigDirTask.kt
@@ -38,6 +38,7 @@ open class PrepareWebpackConfigDirTask : DefaultTask() {
                       },
                     ],
                 });
+                config.devServer = { ...config.devServer, historyApiFallback: true, host: "0.0.0.0" }
             """.trimIndent()
             )
         }