+++ gradle-plugins/src/main/kotlin/versions.kt
@@ -3,5 +3,5 @@ object versions {
     val agp = "4.1.0"
     val nexus_staging = "0.22.0"
     val asoft_builders = "1.2.0"
-    val asoft_mapper = "0.0.2"
+    val asoft_mapper = "0.0.5"
 }