+++ samples/applikation-samples/desktop-applikation/src/main/kotlin/tz/co/asoft/main.kt
@@ -2,7 +2,7 @@ package tz.co.asoft
 
 fun main() {
     val kfg = konfig()
-    val json = Mapper.encodeToString(kfg)
+    val json = Mapper { prettyPrint = true }.encodeToString(kfg)
     println("works: $json")
     val link: String by kfg
     println("It works: Link found was $link")